I use the lower() function to lower all the characters in the string that the generate_username() function returns. I did this because the MainGW api automatically lowers the characters when I create the email, which caused the credentials to not match.
I then made a function called enter_email(). This function gets called after the email has been created on mail.gw. It first clears the "Authorization" header because it was causing reddit to give me an error message. Then it sends a get request to reddit's sign up page and gets the csrf token in the response. It then sends a post request to reddit's api to check if the email is valid. It then returns the response code.

If the response code is 200, it will then move on to the function enter_user_and_password(). This function is what I just started, and it will be responsible for sending a post request of my email, username, and the password to reddit's sign up api.